Added a line "p1=p" in the very beginning.  Without this line, the line p=p1 in the very beginning of program 1 may mess up the current position when we use the new code on an already-being-used motor -- in this case we want to automatically preserve the current position.  This is because when we create the p1 variable if there was not one before, it may take some abritrary value like 0.
The logic is now:
whenever we load the MCode program to a motor -- which should have gone
through the power-up step, no matter there was a "p1" variable saved before
or not, we create it, assign the current position to it (this is new), then
when we instruct the controllor to run program 1, this current position
will be assigned back to the p register, so no change.

Then when there is a power cycle, program 1 will start automatically and
restore the previously saved current position; or if we re-load the MCode
program (p already has a current value), we do p1=p, then program 1 will
do p=p1 -- no change!


git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/pcds_motion/current@5203 ececa60f-9f63-0410-97c7-ea05888eb260